==> 4-Amenity Scores.ipynb <==
Use virtual environment: ors_env

# <font color = 'purple'> Amenity Scores
- Having quantified accessibility of amenities for each apartment, we rank the apartments for easier comparison.
- This ranking will serve as the basis for our apartment search tool

import geopandas as gpd
import pandas as pd
import numpy as np
import math

import os
os.chdir(r"C:\Users\sharo\Documents\Postgrad\My Data Science Portfolio\Transforming the Way We Search For Flats")

## Import Data

amenities = gpd.read_file("Clean Datasets/dist_to_amenities")
amenities.head()

#rename columns that have been abbreviated while saving
amenities.rename(columns={'n_supmkt_4':'n_supmkt_400',
                          'dist_supmk':'dist_supmkt',
                          'n_presch_4':"n_presch_400",
                         'grnarea_40':'grnarea_400'},inplace=True)

#we don't need geometries for this analysis
amenities.drop(columns=['geometry','Latitude','Longitude'],inplace=True)
amenities.head()

#save the file
**STOP! DO NOT OVERWRITE FILE!**
amenities.to_csv("Clean Datasets/amenities_raw.csv",index=False)

## Compute Amenity Scores
- decided to do this for all items so that I can select which ones to use in Tableau
- Aggregated scores will not be calculated here since it's a simple formula. leave it to Tableau

def percentile_ranking(column,small_to_big):
    ranked_column = (column.rank(method='dense', #this ensures blocks with same value get same rank and rank always increases by 1 between groups (1,2,2,3 instead of 1,2,2,4)
                                 pct=True, #express ranking as percentile
                                 na_option='bottom', #NAs occur for min dist items when there weren't any of that amenity within the search radius. these should get lowest rank
                                 ascending=small_to_big)*100).apply(math.ceil) #convert fraction to percentage and round up to nearest 1%
    return(ranked_column)

score_1 = amenities.copy()

#number of supermarkets within 400m
score_1['p_sup400'] = percentile_ranking(score_1['n_supmkt_400'],small_to_big=False) #most supermarkets gets rank1

#distance to nearest supermarket 
score_1['p_supDist'] = percentile_ranking(score_1['dist_supmkt'],small_to_big=True) #shortest distance gets rank 1

#number of food establishments within 400m
score_1['p_fd400'] = percentile_ranking(score_1['n_food_400'], small_to_big=False) #most food est. gets rank 1

#number of mrt stations within 400m
score_1['p_mrt400'] = percentile_ranking(score_1['n_mrt_400m'], small_to_big=False) #most mrt gets rank 1

#distance to nearest mrt
score_1['p_mrtDist'] = percentile_ranking(score_1['dist_mrt'], small_to_big=True) #shortest distance gets rank 1

#distance to CBD
score_1['p_cbdDist'] = percentile_ranking(score_1['dist_cbd'], small_to_big=True) #shortest distance gets rank 1

#number of preschools within 400m
score_1['p_psc400'] = percentile_ranking(score_1['n_presch_400'], small_to_big=False) #most preschools gets rank1

#number of schools within 2km
score_1['p_sch2km'] = percentile_ranking(score_1['n_sch_2km'], small_to_big=False) #most schools gets rank1

#ranking of best primary school within 2km
score_1['p_pri2km'] = percentile_ranking(score_1['prank_2km'], small_to_big=True) #smallest ranking is the best

#green area within 400m
score_1['p_grn400'] = percentile_ranking(score_1['grnarea_400'], small_to_big=False) #highest green area gets rank1

score_1.drop(columns=['n_supmkt_400', 'dist_supmkt', 'n_food_400', 'n_mrt_400m', 'dist_mrt', 
                      'dist_cbd', 'n_presch_400', 'n_sch_2km', 'prank_2km','grnarea_400'], inplace=True)

score_1.head()

## Save

**STOP! DO NOT OVERWRITE FILE!**
score_1.to_csv("Clean Datasets/amenities_scores.csv",index=False)
