==> 5-Preprocessing Resale Prices.ipynb <==
import geopandas as gpd
import pandas as pd
import numpy as np
import math

import os
os.chdir(r"C:\Users\sharo\Documents\Postgrad\My Data Science Portfolio\Transforming the Way We Search For Flats")

## Import Resale Data
- data frm Data.gov.sg
- transactions from Jan 2017-20 Sep 2021

resale = pd.read_csv("HDB Resale Prices/resale-flat-prices-based-on-registration-date-from-jan-2017-onwards.csv",
                    usecols=['block','street_name','flat_type','storey_range','floor_area_sqm','resale_price'])

resale.rename(columns={'block':'blk_no','street_name':'street'},inplace=True)

resale.head()

## Import clean resi-blks dataset

resi_blks = pd.read_csv("Clean Datasets/resi_blks2.csv")
resi_blks=resi_blks.drop(columns=['max_floor_lvl','bldg_contract_town','total_dwelling_units','1room_rental', 
                                  '2room_rental', '3room_rental','other_room_rental', 'n_sold', 
                                  'n_rental', 'n_residents','remaining_lease', 'Latitude', 'Longitude','hdb_town'])
resi_blks.head()

## Add postcodes to resale data

postcode_lookup = resi_blks.loc[:,['blk_no','street','Postcode']]
postcode_lookup.head()

resale_1 = resale.merge(postcode_lookup, how='left', on=['blk_no','street'])
# resale_1['Postcode'] = resale_1['Postcode'].apply(int)

resale_1.head()

#extract the ones which could not be matched
resale_1_nulls = resale_1[resale_1['Postcode'].isnull()]
resale_1_nulls

#fix streetnames in different formats which hinder matching. these were the same ones which we had to fix while doing geocoding for resi-blks

#noticed that almost all the null entries were at St. George's Road. On OneMap website, Saint is spelt out in full. 
resale_1_nulls['street'] = resale_1_nulls['street'].str.replace(pat="ST.", repl="SAINT",case=False,regex=False)

#For the remainin blocks, CENTRAL is spelled as CTRL
resale_1_nulls['street'] = resale_1_nulls['street'].str.replace(pat="CTRL", repl="CENTRAL",case=False,regex=False)

#try to merge again
resale_1_nulls_fixed = resale_1_nulls.drop(columns=['Postcode'])
resale_1_nulls_fixed = resale_1_nulls_fixed.merge(postcode_lookup, how='left', on=['blk_no','street'])
resale_1_nulls_fixed.head()

still_null = resale_1_nulls_fixed[resale_1_nulls_fixed['Postcode'].isnull()]
still_null

Yay!

**Combine the original blks without nulls with the fixed blocks**

resale_1_notnull = resale_1[resale_1['Postcode'].notnull()]
resale_1_notnull.head()

resale_2 = resale_1_notnull.append(resale_1_nulls_fixed,ignore_index=True)
resale_2.head()

#verify no nulls
resale_2.isnull().values.any()

#check number of resale transactions is consistent
len(resale_1)==len(resale_2)

Great job!

### Check % of resi-blks that have at least 1 resale transaction in past 5 years

nblks_haveresale = resi_blks[resi_blks['Postcode'].isin(resale_2['Postcode'])]
nblks_haveresale.head()

pctblks_haveresale = len(nblks_haveresale)/len(resi_blks)*100
pctblks_haveresale

Over 92% of blocks have at least 1 resale transaction in the past 5 years.

## Break down resi-blocks by flat type present in each block

resi_blks.head()

resi_blks2 = resi_blks.copy()
resi_blks2 = resi_blks2.melt(id_vars=['Postcode','blk_no','street'],value_vars=['1room_sold', '2room_sold', '3room_sold',
       '4room_sold', '5room_sold', 'exec_sold', 'multigen_sold', 'studio_apartment_sold'],
                             var_name='flat_type',value_name='n_units')

resi_blks2.sort_values(['Postcode','blk_no','street','flat_type'],inplace=True)

#keep only flat types present in each block
resi_blks2 = resi_blks2[resi_blks2['n_units']>0]
resi_blks2.head(10)

**reformat to facilitate merging**

resi_blks2['flat_type'].value_counts()

resale_2['flat_type'].value_counts()

#translation table for flat type
translate_resiblk = pd.DataFrame.from_dict({'flat_type':['4room_sold','5room_sold','3room_sold','exec_sold',
                                            '2room_sold','studio_apartment_sold','multigen_sold','1room_sold'],
                               'translation': ['4 Rm','5 Rm','3 Rm','Executive', '1 Rm, 2 Rm, Studio Apt',
                                               '1 Rm, 2 Rm, Studio Apt','Multi-Gen','1 Rm, 2 Rm, Studio Apt']})

translate_resiblk

translate_resale = pd.DataFrame.from_dict({'flat_type':['4 ROOM','5 ROOM','3 ROOM','EXECUTIVE',
                                            '2 ROOM','MULTI-GENERATION','1 ROOM'],
                               'translation': ['4 Rm','5 Rm','3 Rm','Executive', '1 Rm, 2 Rm, Studio Apt',
                                               'Multi-Gen','1 Rm, 2 Rm, Studio Apt']})

translate_resale

resi_blks3 = resi_blks2.merge(translate_resiblk,how='left',on='flat_type').drop(columns=['flat_type']).rename(columns={'translation':'flat_type'})
resi_blks3.head()

#check all matched
resi_blks3.isnull().values.any()

resale_3 = resale_2.merge(translate_resale,how='left',on='flat_type').drop(columns=['flat_type']).rename(columns={'translation':'flat_type'})
resale_3.head()

#check all matched
resale_3.isnull().values.any()

## Match median resale price for each flat type in each block
- this involves averaging out between flats in the same floor level band, as well as between bands
- basically we want to see what proporation of blk-flat type combinations have resale data available

resale_4 = resale_3.drop(columns='storey_range').groupby(['Postcode','blk_no','street','flat_type'],as_index=False).median()

resale_4.head()

#merge this with the full list of resi_blks
resi_blks4 = resi_blks3.merge(resale_4,how='left',on=['Postcode','blk_no','street','flat_type'])

resi_blks4.head()

n_matched = len(resi_blks4[resi_blks4['resale_price'].notnull()])

pct_matched=n_matched/len(resi_blks4)*100

pct_matched

Over 80% match, which is decent i think. Reasonable to say that not all blocks would have had resale transactions for all their flat types within the past 5 years

## save files for tableau

**individual resale transactions**

resale_3.head()

#convert postcode to integer. previously was float because of the nulls
resale_3['Postcode'] = resale_3['Postcode'].apply(int)

resale_3.head()

#calculate resale price per square metre
resale_3['resale_price_psm'] = resale_3['resale_price']/resale_3['floor_area_sqm']
resale_3.head()

resale_3.isnull().values.any()

resale_3.sort_values(['Postcode','blk_no','street','flat_type'], inplace=True)

resale_3.head(50)

**STOP DO NOT OVERWRITE!**
resale_3.to_csv("Clean Datasets/resale_Jan2017.csv",index=False)

**median by block**

resale_3_medbyblock = resale_3.drop(columns=['storey_range','floor_area_sqm','resale_price']).groupby(['Postcode','blk_no','street','flat_type'],
                                                                   as_index=False).median()

resale_3_medbyblock = resale_3_medbyblock.pivot(index=['Postcode','blk_no','street'],
                                                columns='flat_type',values='resale_price_psm').reset_index()

resale_3_medbyblock.head()

**STOP DO NOT OVERWRITE!**
resale_3_medbyblock.to_csv("Clean Datasets/resale_Jan2017_BlkMedians.csv",index=False)
