==> 6-Million Dollar Flat List.ipynb <==
Virtual environment: ors_env

import pandas as pd

import os
os.chdir(r"C:\Users\sharo\Documents\Postgrad\My Data Science Portfolio\Transforming the Way We Search For Flats")

mdfs = pd.read_csv("Clean Datasets/resale_Jan2017.csv")
mdfs.head()

#filter million dollar transactions
mdfs_1 = mdfs[(mdfs['resale_price']>=1000000)]

#drop columns we don't need
mdfs_1 = mdfs_1.loc[:,['Postcode','blk_no','street','flat_type']]
mdfs_1.rename(columns={'flat_type':'mdf_type'}, inplace=True)

#remove duplicates
mdfs_1.drop_duplicates(ignore_index=True, inplace=True)

mdfs_1

mdfs_2 = mdfs_1[mdfs_1['mdf_type'].notnull()]

mdfs_2

#save
**STOP DO NOT OVERWRITE!**
mdfs_1.to_csv("Clean Datasets/million-dollar-flats.csv", index=False)
